Update Invoke-HardeningKitty.ps1
Change for comment and verify the function of the result. Results can appear null or empty and cause a bug.

The bug is in Powershell script at line 1327: "You cannot call a method on a null-valued expression."
Details here: "
    + CategoryInfo          : InvalidOperation: (:) [], RuntimeException
    + FullyQualifiedErrorId : InvokeMethodOnNull
"
It can be replicated using the finding list cis_microsoft_windows_server_2019_1809_1.2.1_machine.csv